refactoring things in order to extract UI related code to television-screen
