+++ crates/television-screen/src/help.rs
@@ -2,12 +2,12 @@ use crate::layout::HelpBarLayout;
 use crate::logo::build_logo_paragraph;
 use crate::metadata::build_metadata_table;
 use crate::mode::{mode_color, Mode};
-use color_eyre::eyre::Result;
 use ratatui::layout::Rect;
 use ratatui::prelude::{Color, Style};
 use ratatui::widgets::{Block, BorderType, Borders, Padding, Table};
 use ratatui::Frame;
 use television_channels::channels::UnitChannel;
+use television_utils::metadata::AppMetadata;
 
 pub fn draw_logo_block(f: &mut Frame, area: Rect, color: Color) {
     let logo_block = Block::default()
@@ -27,6 +27,7 @@ fn draw_metadata_block(
     area: Rect,
     mode: Mode,
     current_channel: UnitChannel,
+    app_metadata: &AppMetadata,
 ) {
     let metadata_block = Block::default()
         .borders(Borders::ALL)
@@ -36,7 +37,8 @@ fn draw_metadata_block(
         .style(Style::default());
 
     let metadata_table =
-        build_metadata_table(mode, current_channel).block(metadata_block);
+        build_metadata_table(mode, current_channel, app_metadata)
+            .block(metadata_block);
 
     f.render_widget(metadata_table, area);
 }
@@ -60,9 +62,16 @@ pub fn draw_help_bar(
     current_channel: UnitChannel,
     keymap_table: Table,
     mode: Mode,
+    app_metadata: &AppMetadata,
 ) {
     if let Some(help_bar) = layout {
-        draw_metadata_block(f, help_bar.left, mode, current_channel);
+        draw_metadata_block(
+            f,
+            help_bar.left,
+            mode,
+            current_channel,
+            app_metadata,
+        );
         draw_keymaps_block(f, help_bar.middle, keymap_table);
         draw_logo_block(f, help_bar.right, mode_color(mode));
     }
