+++ crates/television-screen/src/lib.rs
@@ -1,11 +1,13 @@
 pub mod cache;
-mod colors;
+pub mod colors;
 pub mod help;
+pub mod input;
+pub mod keybindings;
 pub mod layout;
 pub mod logo;
 pub mod metadata;
 pub mod mode;
 pub mod preview;
-mod remote_control;
+pub mod remote_control;
 pub mod results;
 pub mod spinner;
