+++ crates/television-screen/src/mode.rs
@@ -1,9 +1,9 @@
 use ratatui::style::Color;
 use serde::{Deserialize, Serialize};
 
-const CHANNEL_COLOR: Color = Color::Indexed(222);
-const REMOTE_CONTROL_COLOR: Color = Color::Indexed(1);
-const SEND_TO_CHANNEL_COLOR: Color = Color::Indexed(105);
+pub const CHANNEL_COLOR: Color = Color::Indexed(222);
+pub const REMOTE_CONTROL_COLOR: Color = Color::Indexed(1);
+pub const SEND_TO_CHANNEL_COLOR: Color = Color::Indexed(105);
 
 pub fn mode_color(mode: Mode) -> Color {
     match mode {
@@ -13,6 +13,7 @@ pub fn mode_color(mode: Mode) -> Color {
     }
 }
 
+// FIXME: Mode shouldn't be in the screen crate
 #[derive(PartialEq, Copy, Clone, Hash, Eq, Debug, Serialize, Deserialize)]
 pub enum Mode {
     Channel,
