+++ crates/television-screen/src/remote_control.rs
@@ -1,8 +1,10 @@
+use std::collections::HashMap;
+
 use crate::logo::build_remote_logo_paragraph;
-use crate::mode::{mode_color, Mode};
+use crate::mode::REMOTE_CONTROL_COLOR;
 use crate::results::build_results_list;
-use television_channels::channels::remote_control::RemoteControl;
-use tv::television::Television;
+use television_channels::entry::Entry;
+use television_utils::input::Input;
 
 use crate::colors::{ResultsListColors, BORDER_COLOR, DEFAULT_INPUT_FG};
 use color_eyre::eyre::Result;
@@ -11,17 +13,18 @@ use ratatui::prelude::Style;
 use ratatui::style::{Color, Stylize};
 use ratatui::text::{Line, Span};
 use ratatui::widgets::{
-    Block, BorderType, Borders, ListDirection, Padding, Paragraph,
+    Block, BorderType, Borders, ListDirection, ListState, Padding, Paragraph,
 };
 use ratatui::Frame;
-use television_channels::channels::OnAir;
 
-impl Television {}
 pub fn draw_remote_control(
-    &mut self,
     f: &mut Frame,
     rect: Rect,
-    mode: Mode,
+    entries: &[Entry],
+    use_nerd_font_icons: bool,
+    picker_state: &mut ListState,
+    input_state: &mut Input,
+    icon_color_cache: &mut HashMap<String, Color>,
 ) -> Result<()> {
     let layout = Layout::default()
         .direction(Direction::Vertical)
@@ -34,18 +37,27 @@ pub fn draw_remote_control(
             .as_ref(),
         )
         .split(rect);
-    self.draw_rc_channels(f, &layout[0])?;
-    self.draw_rc_input(f, &layout[1])?;
-    draw_rc_logo(f, layout[2], mode_color(mode));
+    draw_rc_channels(
+        f,
+        layout[0],
+        entries,
+        use_nerd_font_icons,
+        picker_state,
+        icon_color_cache,
+    );
+    draw_rc_input(f, layout[1], input_state)?;
+    draw_rc_logo(f, layout[2]);
     Ok(())
 }
 
 fn draw_rc_channels(
-    &mut self,
     f: &mut Frame,
-    area: &Rect,
-    remote: RemoteControl,
-) -> Result<()> {
+    area: Rect,
+    entries: &[Entry],
+    use_nerd_font_icons: bool,
+    picker_state: &mut ListState,
+    icon_color_cache: &mut HashMap<String, Color>,
+) {
     let rc_block = Block::default()
         .borders(Borders::ALL)
         .border_type(BorderType::Rounded)
@@ -53,33 +65,21 @@ fn draw_rc_channels(
         .style(Style::default())
         .padding(Padding::right(1));
 
-    let result_count = self.remote_control.result_count();
-    if result_count > 0 && self.rc_picker.selected().is_none() {
-        self.rc_picker.select(Some(0));
-        self.rc_picker.relative_select(Some(0));
-    }
-
-    let entries = self.remote_control.results(
-        area.height.saturating_sub(2).into(),
-        u32::try_from(self.rc_picker.offset())?,
-    );
-
     let channel_list = build_results_list(
         rc_block,
-        &entries,
+        entries,
         ListDirection::TopToBottom,
         Some(
-            ResultsListColors::default().result_name_fg(mode_color(self.mode)),
+            ResultsListColors::default().result_name_fg(REMOTE_CONTROL_COLOR),
         ),
-        self.config.ui.use_nerd_font_icons,
-        &mut self.icon_color_cache,
+        use_nerd_font_icons,
+        icon_color_cache,
     );
 
-    f.render_stateful_widget(channel_list, *area, &mut self.rc_picker.state);
-    Ok(())
+    f.render_stateful_widget(channel_list, area, picker_state);
 }
 
-fn draw_rc_input(&mut self, f: &mut Frame, area: &Rect) -> Result<()> {
+fn draw_rc_input(f: &mut Frame, area: Rect, input: &mut Input) -> Result<()> {
     let input_block = Block::default()
         .title_top(Line::from("Remote Control").alignment(Alignment::Center))
         .borders(Borders::ALL)
@@ -87,9 +87,9 @@ fn draw_rc_input(&mut self, f: &mut Frame, area: &Rect) -> Result<()> {
         .border_style(Style::default().fg(BORDER_COLOR))
         .style(Style::default());
 
-    let input_block_inner = input_block.inner(*area);
+    let input_block_inner = input_block.inner(area);
 
-    f.render_widget(input_block, *area);
+    f.render_widget(input_block, area);
 
     // split input block into 2 parts: prompt symbol, input
     let inner_input_chunks = Layout::default()
@@ -113,29 +113,28 @@ fn draw_rc_input(&mut self, f: &mut Frame, area: &Rect) -> Result<()> {
     let interactive_input_block = Block::default();
     // keep 2 for borders and 1 for cursor
     let width = inner_input_chunks[1].width.max(3) - 3;
-    let scroll = self.rc_picker.input.visual_scroll(width as usize);
-    let input = Paragraph::new(self.rc_picker.input.value())
+    let scroll = input.visual_scroll(width as usize);
+    let input_paragraph = Paragraph::new(input.value())
         .scroll((0, u16::try_from(scroll)?))
         .block(interactive_input_block)
         .style(Style::default().fg(DEFAULT_INPUT_FG).bold().italic())
         .alignment(Alignment::Left);
-    f.render_widget(input, inner_input_chunks[1]);
+    f.render_widget(input_paragraph, inner_input_chunks[1]);
 
     // Make the cursor visible and ask tui-rs to put it at the
     // specified coordinates after rendering
     f.set_cursor_position((
         // Put cursor past the end of the input text
         inner_input_chunks[1].x
-            + u16::try_from(
-                self.rc_picker.input.visual_cursor().max(scroll) - scroll,
-            )?,
+            + u16::try_from(input.visual_cursor().max(scroll) - scroll)?,
         // Move one line down, from the border to the input line
         inner_input_chunks[1].y,
     ));
     Ok(())
 }
-fn draw_rc_logo(f: &mut Frame, area: Rect, color: Color) {
-    let logo_block = Block::default().style(Style::default().fg(color));
+fn draw_rc_logo(f: &mut Frame, area: Rect) {
+    let logo_block =
+        Block::default().style(Style::default().fg(REMOTE_CONTROL_COLOR));
 
     let logo_paragraph = build_remote_logo_paragraph()
         .alignment(Alignment::Center)
