+++ crates/television-utils/Cargo.toml
@@ -21,6 +21,7 @@ bat = { version = "0.24.0", default-features = false, features = ["regex-onig"] 
 directories = "5.0.1"
 syntect = "5.2.0"
 gag = "1.0.0"
+unicode-width = "0.2.0"
 
 [target.'cfg(windows)'.dependencies]
 winapi-util = "0.1.9"
