+++ crates/television/main.rs
@@ -21,13 +21,12 @@ pub mod config;
 pub mod errors;
 pub mod event;
 pub mod input;
+pub mod keymap;
 pub mod logging;
 pub mod picker;
 pub mod render;
 pub mod television;
 pub mod tui;
-pub mod ui;
-mod ui;
 
 #[tokio::main(flavor = "multi_thread")]
 async fn main() -> Result<()> {
