+++ crates/television/television.rs
@@ -1,7 +1,8 @@
-use crate::input::actions::InputActionHandler;
+use crate::config::KeyBindings;
+use crate::input::convert_action_to_input_request;
 use crate::picker::Picker;
 use crate::{action::Action, config::Config};
-use crate::{app::Keymap, cable::load_cable_channels};
+use crate::{cable::load_cable_channels, keymap::Keymap};
 use color_eyre::Result;
 use copypasta::{ClipboardContext, ClipboardProvider};
 use ratatui::{layout::Rect, style::Color, Frame};
@@ -15,10 +16,19 @@ use television_channels::entry::{Entry, ENTRY_PLACEHOLDER};
 use television_previewers::previewers::Previewer;
 use television_screen::cache::RenderedPreviewCache;
 use television_screen::help::draw_help_bar;
+use television_screen::input::draw_input_box;
+use television_screen::keybindings::{
+    build_keybindings_table, DisplayableAction, DisplayableKeybindings,
+};
 use television_screen::layout::{Dimensions, InputPosition, Layout};
 use television_screen::mode::Mode;
-use television_screen::preview::draw_preview_content_block;
+use television_screen::preview::{
+    draw_preview_content_block, draw_preview_title_block,
+};
+use television_screen::remote_control::draw_remote_control;
+use television_screen::results::draw_results_list;
 use television_screen::spinner::{Spinner, SpinnerState};
+use television_utils::metadata::{AppMetadata, BuildMetadata};
 use television_utils::strings::EMPTY_STRING;
 use tokio::sync::mpsc::UnboundedSender;
 
@@ -41,6 +51,7 @@ pub struct Television {
     pub rendered_preview_cache: Arc<Mutex<RenderedPreviewCache<'static>>>,
     pub(crate) spinner: Spinner,
     pub(crate) spinner_state: SpinnerState,
+    pub app_metadata: AppMetadata,
 }
 
 impl Television {
@@ -54,6 +65,18 @@ impl Television {
         let keymap = Keymap::from(&config.keybindings);
         let builtin_channels = load_builtin_channels();
         let cable_channels = load_cable_channels().unwrap_or_default();
+        let app_metadata = AppMetadata::new(
+            env!("CARGO_PKG_VERSION").to_string(),
+            BuildMetadata::new(
+                env!("VERGEN_RUSTC_SEMVER").to_string(),
+                env!("VERGEN_BUILD_DATE").to_string(),
+                env!("VERGEN_CARGO_TARGET_TRIPLE").to_string(),
+            ),
+            std::env::current_dir()
+                .expect("Could not get current directory")
+                .to_string_lossy()
+                .to_string(),
+        );
 
         channel.find(EMPTY_STRING);
         let spinner = Spinner::default();
@@ -80,6 +103,7 @@ impl Television {
             )),
             spinner,
             spinner_state: SpinnerState::from(&spinner),
+            app_metadata,
         }
     }
 
@@ -167,7 +191,7 @@ impl Television {
         match self.mode {
             Mode::Channel => self.results_picker.reset_selection(),
             Mode::RemoteControl | Mode::SendToChannel => {
-                self.rc_picker.reset_selection()
+                self.rc_picker.reset_selection();
             }
         }
     }
@@ -176,7 +200,7 @@ impl Television {
         match self.mode {
             Mode::Channel => self.results_picker.reset_input(),
             Mode::RemoteControl | Mode::SendToChannel => {
-                self.rc_picker.reset_input()
+                self.rc_picker.reset_input();
             }
         }
     }
@@ -241,7 +265,8 @@ impl Television {
                         &mut self.rc_picker.input
                     }
                 };
-                input.handle_action(&action);
+                input
+                    .handle(convert_action_to_input_request(&action).unwrap());
                 match action {
                     Action::AddInputChar(_)
                     | Action::DeletePrevChar
@@ -374,18 +399,50 @@ impl Television {
             f,
             &layout.help_bar,
             self.current_channel(),
-            self.build_keymap_table()?,
+            build_keybindings_table(
+                &self.config.keybindings.to_displayable(),
+                self.mode,
+            ),
             self.mode,
+            &self.app_metadata,
         );
 
-        self.results_area_height = u32::from(layout.results.height - 2); // 2 for the borders
+        self.results_area_height =
+            u32::from(layout.results.height.saturating_sub(2)); // 2 for the borders
         self.preview_pane_height = layout.preview_window.height;
 
         // results list
-        self.draw_results_list(f, layout.results)?;
+        let result_count = self.channel.result_count();
+        if result_count > 0 && self.results_picker.selected().is_none() {
+            self.results_picker.select(Some(0));
+            self.results_picker.relative_select(Some(0));
+        }
+        let entries = self.channel.results(
+            self.results_area_height,
+            u32::try_from(self.results_picker.offset())?,
+        );
+        draw_results_list(
+            f,
+            layout.results,
+            &entries,
+            &mut self.results_picker.relative_state,
+            self.config.ui.input_bar_position,
+            self.config.ui.use_nerd_font_icons,
+            &mut self.icon_color_cache,
+        )?;
 
         // input box
-        self.draw_input_box(f, layout.input)?;
+        draw_input_box(
+            f,
+            layout.input,
+            result_count,
+            self.channel.total_count(),
+            &mut self.results_picker.input,
+            &mut self.results_picker.state,
+            self.channel.running(),
+            &self.spinner,
+            &mut self.spinner_state,
+        )?;
 
         let selected_entry = self
             .get_selected_entry(Some(Mode::Channel))
@@ -394,7 +451,7 @@ impl Television {
 
         // preview title
         self.current_preview_total_lines = preview.total_lines();
-        self.draw_preview_title_block(
+        draw_preview_title_block(
             f,
             layout.preview_title,
             &preview,
@@ -404,8 +461,9 @@ impl Television {
         // preview content
         // initialize preview scroll
         self.maybe_init_preview_scroll(
-            entry.line_number.map(|l| u16::try_from(l).unwrap_or(0)),
-            // FIXME: adjust this if necessary
+            selected_entry
+                .line_number
+                .map(|l| u16::try_from(l).unwrap_or(0)),
             layout.preview_window.height,
         );
         draw_preview_content_block(
@@ -419,7 +477,25 @@ impl Television {
 
         // remote control
         if matches!(self.mode, Mode::RemoteControl | Mode::SendToChannel) {
-            self.draw_remote_control(f, layout.remote_control.unwrap())?;
+            // NOTE: this should be done in the `update` method
+            let result_count = self.remote_control.result_count();
+            if result_count > 0 && self.rc_picker.selected().is_none() {
+                self.rc_picker.select(Some(0));
+                self.rc_picker.relative_select(Some(0));
+            }
+            let entries = self.remote_control.results(
+                area.height.saturating_sub(2).into(),
+                u32::try_from(self.rc_picker.offset())?,
+            );
+            draw_remote_control(
+                f,
+                layout.remote_control.unwrap(),
+                &entries,
+                self.config.ui.use_nerd_font_icons,
+                &mut self.rc_picker.state,
+                &mut self.rc_picker.input,
+                &mut self.icon_color_cache,
+            )?;
         }
         Ok(())
     }
@@ -434,4 +510,140 @@ impl Television {
                 Some(target_line.unwrap_or(0).saturating_sub(height / 3));
         }
     }
+}
+
+impl KeyBindings {
+    pub fn to_displayable(&self) -> HashMap<Mode, DisplayableKeybindings> {
+        // channel mode keybindings
+        let channel_bindings: HashMap<DisplayableAction, Vec<String>> =
+            HashMap::from_iter(vec![
+                (
+                    DisplayableAction::ResultsNavigation,
+                    serialized_keys_for_actions(
+                        self,
+                        &[
+                            Action::SelectPrevEntry,
+                            Action::SelectNextEntry,
+                            Action::SelectPrevPage,
+                            Action::SelectNextPage,
+                        ],
+                    ),
+                ),
+                (
+                    DisplayableAction::PreviewNavigation,
+                    serialized_keys_for_actions(
+                        self,
+                        &[
+                            Action::ScrollPreviewHalfPageUp,
+                            Action::ScrollPreviewHalfPageDown,
+                        ],
+                    ),
+                ),
+                (
+                    DisplayableAction::SelectEntry,
+                    serialized_keys_for_actions(self, &[Action::SelectEntry]),
+                ),
+                (
+                    DisplayableAction::CopyEntryToClipboard,
+                    serialized_keys_for_actions(
+                        self,
+                        &[Action::CopyEntryToClipboard],
+                    ),
+                ),
+                (
+                    DisplayableAction::SendToChannel,
+                    serialized_keys_for_actions(
+                        self,
+                        &[Action::ToggleSendToChannel],
+                    ),
+                ),
+                (
+                    DisplayableAction::ToggleRemoteControl,
+                    serialized_keys_for_actions(
+                        self,
+                        &[Action::ToggleRemoteControl],
+                    ),
+                ),
+                (
+                    DisplayableAction::Quit,
+                    serialized_keys_for_actions(self, &[Action::Quit]),
+                ),
+            ]);
+
+        // remote control mode keybindings
+        let remote_control_bindings: HashMap<DisplayableAction, Vec<String>> =
+            HashMap::from_iter(vec![
+                (
+                    DisplayableAction::ResultsNavigation,
+                    serialized_keys_for_actions(
+                        self,
+                        &[Action::SelectPrevEntry, Action::SelectNextEntry],
+                    ),
+                ),
+                (
+                    DisplayableAction::SelectEntry,
+                    serialized_keys_for_actions(self, &[Action::SelectEntry]),
+                ),
+                (
+                    DisplayableAction::ToggleRemoteControl,
+                    serialized_keys_for_actions(
+                        self,
+                        &[Action::ToggleRemoteControl],
+                    ),
+                ),
+            ]);
+
+        // send to channel mode keybindings
+        let send_to_channel_bindings: HashMap<DisplayableAction, Vec<String>> =
+            HashMap::from_iter(vec![
+                (
+                    DisplayableAction::ResultsNavigation,
+                    serialized_keys_for_actions(
+                        self,
+                        &[Action::SelectPrevEntry, Action::SelectNextEntry],
+                    ),
+                ),
+                (
+                    DisplayableAction::SelectEntry,
+                    serialized_keys_for_actions(self, &[Action::SelectEntry]),
+                ),
+                (
+                    DisplayableAction::Cancel,
+                    serialized_keys_for_actions(
+                        self,
+                        &[Action::ToggleSendToChannel],
+                    ),
+                ),
+            ]);
+
+        HashMap::from_iter(vec![
+            (Mode::Channel, DisplayableKeybindings::new(channel_bindings)),
+            (
+                Mode::RemoteControl,
+                DisplayableKeybindings::new(remote_control_bindings),
+            ),
+            (
+                Mode::SendToChannel,
+                DisplayableKeybindings::new(send_to_channel_bindings),
+            ),
+        ])
+    }
+}
+
+fn serialized_keys_for_actions(
+    keybindings: &KeyBindings,
+    actions: &[Action],
+) -> Vec<String> {
+    actions
+        .iter()
+        .map(|a| {
+            keybindings
+                .get(&Mode::Channel)
+                .unwrap()
+                .get(a)
+                .unwrap()
+                .clone()
+                .to_string()
+        })
+        .collect()
 }
